refactored "info" view; added navbar and footer
