+++ public/index.html
@@ -1,55 +1,35 @@
 <!DOCTYPE html>
 
 <html lang="en">
-  <head>
-    <meta charset="utf-8" />
-    <link rel="icon" href="./favicon.png" />
 
-    <meta
-      name="viewport"
-      content="width=device-width, initial-scale=1"
-    />
-    <meta
-      name="theme-color"
-      content="#000000"
-    />
-    <meta
-      name="description"
-      content="Find out the producers behind your favourite songs."
-    />
+<head>
+  <meta charset="utf-8" />
 
-    <link rel="apple-touch-icon" href="./logo192.png" />
-    <link rel="manifest" href="./site.webmanifest" />
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-EVSTQN3/azprG1Anm3QDgpJLIm9Nao0Yz1ztcQTwFspd3yD65VohhpuuCOmLASjC" crossorigin="anonymous">
+  <link rel="icon" href="./favicon.png" />
 
-    <link rel="stylesheet" href="https://tunebat.com/css/bundle.826f59bc5d5f3c05c29e.css" />
+  <meta name="viewport" content="width=device-width, initial-scale=1" />
+  <meta name="theme-color" content="#000000" />
+  <meta name="description" content="Find out the producers behind your favourite songs." />
 
-    <meta property='og:url' content="." />
-    <meta property='og:type' content="website" />
-    <meta property='og:title' content="WhoProduced" />
-    <meta property="og:description"
-      content="Find the producers behind some of your favourite tracks"
-    />
+  <link rel="apple-touch-icon" href="./logo192.png" />
+  <link rel="manifest" href="./site.webmanifest" />
+  <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/css/bootstrap.min.css" rel="stylesheet"
+    integrity="sha384-EVSTQN3/azprG1Anm3QDgpJLIm9Nao0Yz1ztcQTwFspd3yD65VohhpuuCOmLASjC" crossorigin="anonymous" />
 
-    <title>WhoProduced</title>
+  <meta property='og:url' content="." />
+  <meta property='og:type' content="website" />
+  <meta property='og:title' content="WhoProduced" />
+  <meta property="og:description" content="Find the producers behind some of your favourite tracks" />
 
-    <style>
-      @media (max-width: 991px) {
-        .USeLs {
-          border: none  !important;; 
-          box-sizing: border-box!important;
-          border-radius: 4px!important;
-          margin-right: 2px!important;
-        }
-      }
-    </style>
+  <title>WhoProduced</title>
+</head>
 
-  </head>
-  <body>
-    <noscript>You need to enable JavaScript to run this app.</noscript>
+<body>
+  <noscript>You need to enable JavaScript to run this app.</noscript>
 
-    <div id="root"></div>
-    
-    <script src='../src/index.js'></script>
-  </body>
+  <div id="root"></div>
+
+  <script src='../src/index.js'></script>
+</body>
+
 </html>
