+++ src/Components/Footer.js
@@ -1,0 +1,17 @@
+import React from 'react'
+
+export default function Footer() {
+	return (
+		<div>
+			<footer class="page-footer font-small blue">
+				<div className='text-center'>
+					<div>
+						<a href="https://github.com/foderking/WhoProduced_frontend">
+							<img height='80px' src="./github.png" alt="GitHub Logo" draggable="false" />
+						</a>
+					</div>
+				</div>
+			</footer>
+		</div>
+	)
+}
