+++ src/Components/Nav.js
@@ -1,0 +1,13 @@
+import React from 'react'
+
+export default function Nav() {
+	return (
+		<div>
+			<nav className="bg-dark">
+				<a className="navbar navbar-brand" href="/">
+					<img src="./hades.png" width="30" height="30" className="d-inline-block align-top" alt="" />
+				</a>
+			</nav>
+		</div>
+	)
+}
